added a message to the category if theres no items matching
