+++ listCategoriesHTML/onSale.html
@@ -87,7 +87,6 @@
           </div>
         </div>
         <div class="product-list"></div>
-        <a href="#" class="load-button black underline montserrat">Load more</a>
       </section>
       <div class="stickyfooter">
         <div class="copyright_wrapper">RAINYDAYS 2023 &copy;</div>
